Backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
beanie==1.26.0
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
fastapi==0.115.0
h11==0.14.0
idna==3.10
lazy-model==0.2.0
motor==3.6.0
passlib==1.7.4
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pymongo==4.9.1
python-dotenv==1.0.1
sniffio==1.3.1
sqlparse==0.5.1
starlette==0.38.6
toml==0.10.2
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
